api/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fastapi==0.113.0
fonttools==4.53.1
h11==0.14.0
httptools==0.6.1
idna==3.8
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydantic==2.9.0
pydantic_core==2.23.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
